+++ scripts/post-install.sh
@@ -8,6 +8,14 @@ if [ ! -f www/sites/default/settings.php ]
     echo "default.settings.php copied to settings.php"
 fi
 
+# Prepare the services file for installation
+if [ ! -f www/sites/default/services.yml ]
+  then
+    cp www/sites/default/default.services.yml www/sites/default/services.yml
+    chmod 777 www/sites/default/services.yml
+    echo "default.services.yml copied to services.yml"
+fi
+
 # Prepare the files directory for installation
 if [ ! -d www/sites/default/files ]
   then