+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 ## [Unreleased]
 ### Changed
 - Preparing Drupal 8 version
+- Post install script, copy the services.yml file